cg: add dt_cg_trace
A call to

   dt_cg_trace(dlp, drp, 1, 666, foo);

will trace the value of reg FOO into
/sys/kernel/debug/tracing/trace_pipe when dtrace is compiled with
debugging=yes.  '666' above is a counter value which is also logged in
the resulting trace output, to distinguish trace calls from each other.

A call to

  dt_cg_trace(dlp, drp, 0, 666, foo);

will trace the immediate value of FOO at codegen time.

Under debugging=no this call has no effect and is compiled out.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>