alloca: new faults
We'll use DTRACEFLT_BADSIZE to distinguish in error messages between
out-of-bounds errors in bcopy and the like, and size-too-large errors.
(Without this, it's hard to tell if a, say, 0x12 in the fault value
is a nearly-NULL pointer or a wrong size.)

We'll use DTRACEFLT_INTERR to indicate can-never-happen errors where we
have tested the actual condition already (and emitted a proper error)
followed by a check to placate the verifier (which will always succeed
if the actual condition check failed).  The latter failure happening at
runtime always indicates a bug in DTrace, not in the user's code, so
indicate as much.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>