alloca: deref
After the last commit we can alloca()te, store the returned pointers in
variables and get them back out looking identical to how they were
stored, with suitable verifier bounds.

All that is left is handling dereferences.  We have two distinct sorts
of derefs to deal with: non-lvalue derefs, like *foo, for which the
load is done at the time of dereferencing and the size is known, and
lvalue loads, like foo[10]=bar, where the size stored is not known
until after the deref is complete.  In both cases we must
re-bounds-check: the pointer that we are dealing with will have been
bounds-checked already, one way or another, but it is common to modify
that pointer (usually via addition, sometimes, e.g. in lvalues,
compiler-generated addition), and we must bounds-check again after that
to make sure it is still in bounds.

For non-lvalue derefs this is quite easy: adjust dt_cg_load so it tells
us the size of load it's generating an op for, and add a bounds check to
the DT_TOK_DEREF case in dt_cg_node.  We do the actual bounds-checking
using a new dt_cg_check_scratch_bounds, which handles the annoying
getting-the-scratch-base part and then calls the bounds checker we
already wrote, dt_cg_check_bounds (exploiting its previously unused
ability to bounds-check not just offsets but pointers too).

For lvalue derefs this is trickier: at DT_TOK_DEREF codegen time we have
no idea what size of store will be carried out, and in fact dt_cg_node
generates no code at all to deref an lvalue store.  The deref is carried
out in dt_cg_store, which has to grow a special case for a store to a
writable lvalue node whose child is a deref with the DT_NF_ALLOCA flag
turned on.  (This combination is generated by the compiler for lvalue
stores, so is completely stereotyped and safe to rely on).  When this
case fires we can do a bounds-check just like the one done for reads.

We can also add a lot of tests and flip off XFAIL for a few pre-existing
ones, now that alloca, and variable storage, and dereferencing all work.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>