alloca: bcopy
bcopy can mostly use our existing bounds-checking machinery, but not
entirely: we also need to handle the 'not in scratch' condition of the
src argument (a condition I'd be happy to drop, frankly, if
bpf_probe_read can handle overlapping copies: it doesn't gain us much
and adds a lot of complexity).  This requires a whole new
bounds-checking function, since it's not an inverse of 'in scratch'; in
particular, both 'not in scratch' and 'in scratch' should reject copies
that overlap the edges of scratch space, and 'not in scratch' has to
handle size overflows, copies that wrap around the edge of the address
space and craziness like that.

This is all complicated a bit because although the bounds checker is
mostly being passed map_values, it must do most bounds-checking on
scalars, because the verifier ignores most bounds checking when either
arg is non-scalar.  To help with this, we add a scratch DMST_SCALARIZER
value ot the machine state, which is just a uint64_t you can round-trip
through to turn pointer-like values into scalars.  (At the end of
bounds-checking, we turn it back into a map_value again via the usual
addition-to-scratchbot approach.)

We add a bunch of tests for all this too.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>