alloca: support null pointers
Variables used to store pointers obtained from alloca() cannot be reused
for any other purpose -- but there is one purpose people are almost
certain to want, which is assigning NULL to them.  You don't want that
to elicit an error message!

Making this work is quite tricky.  We can't just stuff a 0 into
variables and load it, because that would lead to us subtracting
scratchbot from it and then adding scratchbot back afterwards, giving
null pointers a huge random value in the variables.  So we detect
the specific case of var = NULL at codegen time by literal matching the
parse tree and turn it into a store of a sentinel value,
DT_CG_ALLOCA_NULLPTR (~(1<<30)).  We can then detect that at load time
and turn it back into a NULL.

But that's not enough.  We might say "it's up to the user to not
dereference it", but the verifier forces us to ensure that this can't
happen.  We can't leave a 0 in there either because it's not a pointer
value and the verifier's tracking of what happens if a 0 is dereferenced
hits map_value-versus-literal problems: since every *other* value we
can hold in an alloca parse tree node is a pointer, we should arrange
for this to be one too.

But a pointer to what?  It can't be a pointer to scratch space, and it
certainly can't be a pointer to NULL -- that's not a map_value!  So we
introduce a new "null pointer space", in the shape of a one-byte BPF
array, and set null pointers to point at that at load time.  Then it's
just a matter of having the bounds-checkers check for both 0 and the
null pointer space value and raise suitable out-of-bounds errors in both
cases.  (As usual, in-scratch and out-of-scratch bounds checkers
diverge: *both* consider null pointers out of bounds, but the in-scratch
checker doesn't need to do anything to detect that because null pointers
are outside scratch space anyway.  This means null pointer detection
adds very little code except to bcopy.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>