test: Do not return 1 for err.*.x checks
We can mark tests as expected to fail.  In fact, with .x files,
we can mark XFAIL based on a runtime check.

For err.* tests, however, we are expected to fail anyhow.  So,
having an err.*.x return 1 is at least confusing... certainly
for the runtest.sh script.

In practice, err.*.x files usually return only 0 or 2 -- that is,
we might elect to skip the test.

The sole exception is err.unloaded_var.x, which checks for
/lib/modules/$(uname -r)/kernel/fs/gfs2/gfs2.ko*.  If this module
is missing, the test is supposed to fail, but as it is err.* it is
expected to fail anyhow.  So the counterintuitive result is that
if the module is missing, the test produces XPASS.

Change this err.*.x file to skip the test if the module is missing.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>