Ignore clauses: some clauses are impossible regardless of uprp
In ignore_clauses, for an underlying probe uprp, we try to
decide if we can safely ignore clause n.

Meanwhile, for some clauses, the probe description tells us the
clause will not be called for any USDT probe, regardless of the
underlying probe.  For example, "syscall::write:" can safely be
ignored, for all uprp.

Add clause flags to track a statement's status.  Specifically,
a flag may indicate that a statement can be excluded for every
possible USDT probe or included because it might be needed for
some possible USDT probe.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>