test: Clean up zero-size speculation + chill()
The chill() action is not yet supported, but it also is not needed
for the test.  Therefore, remove the chill() from the legacy test.

But we do not want to diminish test coverage.  Therefore, make a
copy of the test that keeps the chill (and annotates its XFAIL
more clearly).

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>