test: testsuite $dt_flags and linker flags problems
The testsuite passes in a $dtrace environment variable: this must
be an ELF executable because some tests use it in hashbang lines,
so necessary flags are passed in in a $dt_flags variable which
must be passed to every dtrace invocation.  These flags are largely
path adjustments (-I etc) pointing DTrace to the in-source-tree
testsuite or the out-of-tree SDT headers and such things: as such,
it's unfortunate that a great many invocations were missed and that
those invocations were largely things like -G that actually *use* -I.

It also passes in a $test_cppflags variable which is meant to be
used for C preprocessor invocations, but many tests which should
use it don't use it, leading to the possibility (and, these days,
the likelihood) of picking up SystemTap <sys/sdt.h> instead, or
simply a compilation error if an installed testsuite is run.

This uncovers another bug: the -I flags passed in in all the
above cases are relative paths, so they are wrong for any tests
that cd, such as almost any test that does any compiles (which
all cd into directories under the $tmpdir). Make them absolute.

runtest.sh also populates a test_ldflags containing -L arguments
for libraries in the build tree, but this is never exported or
used by any tests: pass that in too, and change linker invocations
to use it instead.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>