tests: repair accidentally deleted tst.pidprobes.sh
My intent when writing the earlier tests was to have
test/unittest/usdt/pidprobes.sh as a sort of shell script library
called by tst.pidprobes.sh, tst.pidargs.sh and tst.pidargmap.sh.
Alas, *.sh is executed as a test by runtest.sh, so
pidprobes.sh was treated as a test, not a library.  So rename it
back to what it was called before and make sure it works when
invoked directly by runtest.sh, but also let it be called by
the other two tests to test other corners of this.

(Also: apply some small review comments that never made it
before the previous patch hit dev.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>