tp: clean up the API
All functions that can operate on a tracepoint object had a convenience
function for probes that are purely tracepoint-based, except the setting
and getting of the (event of BTF) id.  We now have dt_tp_probe_* and
dt_tp_* variants for that.

Also renamed event_id/event_fd to be id/fd.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>