Improve dt_pid_create_usdt_probes() error handling
In one run of the test suite, roughly 90 consecutive tests failed.
The exact circumstances are difficult to reconstruct, but basically
they complained with messages like:

    dtrace: failed to compile script ...: failed to grab process 23132

The same pid was always indicated.  Then, the errors stopped.

Apparently, the problem was trying "dtrace -s foo.d" where the script
has a BEGIN (or END) probe.  Then we get to

    cmd/dtrace.c compile_file
    ->  libdtrace/dt_cc.c dt_program_compile
    ->  libdtrace/dt_cc.c dt_compile
    ->  libdtrace/dt_cc.c dt_compile_one_clause
    ->  libdtrace/dt_cc.c dt_setcontext
    ->  libdtrace/dt_pid.c dt_pid_create_usdt_probes

We look for processes that might have USDT probes.  Since the provider
description is blank, we check every process in .../run/dtrace.  It
turned out, that pid 23132 could not be locked.  This sent an error
back up the call stack.

We do not know why process 23132 could not be locked, but having dtrace
fail under these circumstances (using BEGIN or END probe) seems severe.

Change dt_pid_create_usdt_probes() error handling.  Even if there is a
problem with some USDT processes, report success anyhow if there was a
glob pid specification.  If, on the other hand, a pid was specifically
requested, then a problem with that pid results in an error.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>